Updated Tests
The tests has been updated. Due to the flag checking if an item exists, it caused a few errors but the code has been refactored for the main test to see if a value has been set.

Raul on the Udacity forums kindly shared a PR to a fix for the issue with the provided unit test when buying the token. This was a JavaScript based issue with it being unable to handle such HUGE numbers! Using the Web3 utils library and the BN function, it allows us to handle such large numbers safely